+++ lib/src/parser/character/range.dart
@@ -5,11 +5,11 @@ import 'predicate.dart';
 
 /// Returns a parser that accepts any character in the range
 /// between [start] and [stop].
-Parser<String> range(Object start, Object stop, [String? message]) =>
+Parser<String> range(String start, String stop, [String? message]) =>
     CharacterParser(
         RangeCharPredicate(toCharCode(start), toCharCode(stop)),
         message ??
-            '${toReadableString(start)}..${toReadableString(stop)} expected');
+            '[${toReadableString(start)}-${toReadableString(stop)}] expected');
 
 class RangeCharPredicate implements CharacterPredicate {
   RangeCharPredicate(this.start, this.stop) {